Update Schema datatype

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, DateTime
+from sqlalchemy import Column, String, DateTime, Numeric
 from sqlalchemy.sql import func
 from app.database import Base
 
@@ -8,7 +8,7 @@ class Transaction(Base):
     transaction_id = Column(String, primary_key=True, index=True)
     source_account = Column(String, nullable=False)
     destination_account = Column(String, nullable=False)
-    amount = Column(Integer, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     currency = Column(String, nullable=False)
     status = Column(String, nullable=False)
     error_message = Column(String, nullable=True)
